+++ hooks_util/src/profile.rs
@@ -138,8 +138,8 @@ impl debug::Inspect for Node {
                 debug::Vars::Leaf(format!(
                     "{:3.2}% {:>4.2}ms/call {:.2}Hz",
                     percent,
-                    duration_sum_secs * 1000.0 / (self.num_calls as f64),
-                    self.num_calls as f64 / root_duration_sum_secs
+                    duration_sum_secs * 1000.0 / (self.num_calls as f32),
+                    self.num_calls as f32 / root_duration_sum_secs
                 )),
             ),
         ];
